Split IActiveSession interface to extract ILocalSession interface.
Introduce ILocalSession Interface.
Move some methods and properties from IActiveSession to ILocalSesson.
Add BaseId property for ILocalSession interface and implement it in ActiveSession class. Add NullLocalSessionObject.
Those changes were made in preparation for introducing local session - a simplified version of an active session that will have a different lifetime - same as of an ASP.NET Core session and one-to-one relation with it.