Make a dictionary of base information in the ActiveSessionStore class.
(cherry-picked from commit "Make list of environt providers..." for the "environment" feature branch.)
The dictionary is intended to contain common information for all active sessions based on the same LocalSession (having the same BaseId that is the key) such as reference to this LocalSession.
The names of the dictionary field and the dictionary data item type are rather confusing but they are diliberately chosen to coinside with the names in the branch from wich they were imported.